+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 #Changelog
 
+### 0.0.1 (04/30/2016)
+
+**Features**
+
+- Switched to a cli based runner
+- Added fallout4 test files and specs
+- fancier build scripts, spinners and such
+
+**Bugs**
+
+- Can now run tests with .babelrc and rollup with separate config
+
 ### 0.0.0 (04/06/2016)
 
 **Features**